Saving new user preference in API
